增加成员添加功能，权限控制

+++ src/main/java/cn/wzy/entity/ProjectDocumentInfo.java
@@ -40,7 +40,7 @@ public class ProjectDocumentInfo implements Serializable {
 
     //@ApiModelProperty(value = "文件大小")
     @TableField("size")
-    private Long size;
+    private Double size;
 
     //@ApiModelProperty(value = "组织id")
     @TableField("upload_user_id")
